Move wind.ts from formulas to phenomena directory
