+++ src/index.ts
@@ -1,7 +1,7 @@
 import * as humidity from "./formulas/humidity";
 import * as temperature from "./formulas/temperature";
 import * as pressure from "./formulas/pressure";
-import * as wind from "./formulas/wind";
+import * as wind from "./phenomena/wind";
 import * as airDensity from "./formulas/airDensity";
 
 import * as beaufort from "./scales/beaufort";
